refactor: update table
